requirements.txt
streamlit==1.32.0
langchain==0.1.12
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.1
groq==0.4.2
langchain-ollama==0.0.7


https://github.com/lacerdafh/Terapia-intensiva.git